requirements.txt
# Momentum Watchlist Dependencies

# Yahoo Finance (FREE, no API key needed!)
yfinance>=0.2.40

# Alpaca Markets SDK (optional, for real-time data)
alpaca-py>=0.13.0

# Email via SendGrid
sendgrid>=6.10.0

# Data processing
pandas>=2.0.0

# Market calendar
exchange-calendars>=4.5.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (for Alpaca screener API)
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking (optional, for development)
# mypy>=1.0.0
# types-requests>=2.31.0

